documents/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LightRAG 의존성
lightrag>=0.0.1
nano-vectordb>=0.0.1
ollama>=0.1.7
neo4j>=5.14.0
tenacity>=8.2.0
networkx>=3.0
numpy>=1.24.0
scipy>=1.10.0

# 파일 처리
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
markdown==3.5.1
python-pptx==0.6.23

# 비동기 처리
aiofiles==23.2.1
httpx==0.25.2

# 유틸리티
requests==2.31.0
typing-extensions==4.8.0

# 로깅 및 모니터링
structlog==23.2.0
rich==13.7.0

# 보안
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
bcrypt==4.0.1

# 테스트 (개발용)
pytest==7.4.3
pytest-asyncio==0.21.1